+++ package.json
@@ -35,14 +35,10 @@
     "babel-plugin-lodash": "^3.3.4",
     "browser-env": "^3.3.0",
     "eslint": "^7.17.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
     "lodash": "^4.17.20",
     "node-fetch": "^3.0.0-beta.9",
     "nyc": "^15.1.0",
-    "react-test-renderer": "^17.0.1",
     "source-map-support": "^0.5.19",
-    "start-server-and-test": "^1.11.7",
-    "terser-webpack-plugin": "^5.1.1",
     "ts-loader": "^8.0.14",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3",
